Refactor: types: move out currency types from data slice
